Remove the `m` parameter from TransformerResult.
Turns out that letting transformers talk about the monad in their
results leads to unexpected behavior of some functions. Running a
transformer should really mean getting it's effect fully evaluated. This
means that running ListT will evaluate all of the results which is not
always desireable, but it's much more predictable than the alternative.